+++ wimoor-admin/src/main/java/com/wimoor/admin/service/impl/SysRoleServiceImpl.java
@@ -34,7 +34,7 @@ public class SysRoleServiceImpl extends ServiceImpl<SysRoleMapper, SysRole> impl
     @Override
     public boolean delete(List<BigInteger> ids) {
         Optional.ofNullable(ids).orElse(new ArrayList<>()).forEach(id -> {
-            int count = iSysUserRoleService.count(new LambdaQueryWrapper<SysUserRole>().eq(SysUserRole::getRoleId, id));
+            long count = iSysUserRoleService.count(new LambdaQueryWrapper<SysUserRole>().eq(SysUserRole::getRoleId, id));
             Assert.isTrue(count <= 0, "该角色已分配用户，无法删除");
             iSysRoleMenuService.remove(new LambdaQueryWrapper<SysRoleMenu>().eq(SysRoleMenu::getRoleId, id));
             iSysRolePermissionService.remove(new LambdaQueryWrapper<SysRolePermission>().eq(SysRolePermission::getRoleId, id));
